ref(streams): Move `ConsumerError` to streams.consumer
