+++ snuba/utils/streams/kafka.py
@@ -21,15 +21,8 @@ from confluent_kafka import TopicPartition as ConfluentTopicPartition
 
 from snuba.utils.retries import NoRetryPolicy, RetryPolicy
 from snuba.utils.streams.codecs import Codec
-from snuba.utils.streams.consumer import Consumer
-from snuba.utils.streams.types import (
-    ConsumerError,
-    EndOfPartition,
-    Message,
-    Partition,
-    Topic,
-    TPayload,
-)
+from snuba.utils.streams.consumer import Consumer, ConsumerError, EndOfPartition
+from snuba.utils.streams.types import Message, Partition, Topic, TPayload
 
 logger = logging.getLogger(__name__)
