+++ snuba/utils/streams/types.py
@@ -41,26 +41,3 @@ class Message(Generic[TPayload]):
 
     def get_next_offset(self) -> int:
         return self.offset + 1
-
-
-class ConsumerError(Exception):
-    """
-    Base class for exceptions that are raised during consumption.
-
-    Subclasses may extend this class to disambiguate errors that are specific
-    to their implementation.
-    """
-
-
-class EndOfPartition(ConsumerError):
-    """
-    Raised when there are no more messages to consume from the partition.
-    """
-
-    def __init__(self, partition: Partition, offset: int):
-        # The partition that the consumer has reached the end of.
-        self.partition = partition
-
-        # The next unconsumed offset in the partition (where there is currently
-        # no message.)
-        self.offset = offset
