+++ tests/utils/streams/test_kafka.py
@@ -5,6 +5,7 @@ import pytest
 from confluent_kafka import Producer as ConfluentProducer
 from confluent_kafka.admin import AdminClient, NewTopic
 
+from snuba.utils.streams.consumer import Consumer, ConsumerError, EndOfPartition
 from snuba.utils.streams.codecs import PassthroughCodec
 from snuba.utils.streams.kafka import (
     Commit,
@@ -15,8 +16,6 @@ from snuba.utils.streams.kafka import (
     as_kafka_configuration_bool,
 )
 from snuba.utils.streams.types import (
-    ConsumerError,
-    EndOfPartition,
     Message,
     Partition,
     Topic,
@@ -45,7 +44,7 @@ def topic() -> Iterator[Topic]:
 
 
 def test_consumer_backend(topic: Topic) -> None:
-    def build_consumer() -> KafkaConsumer[KafkaPayload]:
+    def build_consumer() -> Consumer[KafkaPayload]:
         codec: PassthroughCodec[KafkaPayload] = PassthroughCodec()
         return KafkaConsumer(
             {
@@ -194,7 +193,7 @@ def test_auto_offset_reset_earliest(topic: Topic) -> None:
     assert producer.flush(5.0) == 0
 
     codec: PassthroughCodec[KafkaPayload] = PassthroughCodec()
-    consumer = KafkaConsumer(
+    consumer: Consumer[KafkaPayload] = KafkaConsumer(
         {
             **configuration,
             "auto.offset.reset": "earliest",
@@ -222,7 +221,7 @@ def test_auto_offset_reset_latest(topic: Topic) -> None:
     assert producer.flush(5.0) == 0
 
     codec: PassthroughCodec[KafkaPayload] = PassthroughCodec()
-    consumer = KafkaConsumer(
+    consumer: Consumer[KafkaPayload] = KafkaConsumer(
         {
             **configuration,
             "auto.offset.reset": "latest",
@@ -254,7 +253,7 @@ def test_auto_offset_reset_error(topic: Topic) -> None:
     assert producer.flush(5.0) == 0
 
     codec: PassthroughCodec[KafkaPayload] = PassthroughCodec()
-    consumer = KafkaConsumer(
+    consumer: Consumer[KafkaPayload] = KafkaConsumer(
         {
             **configuration,
             "auto.offset.reset": "error",
@@ -287,7 +286,7 @@ def test_commit_log_consumer(topic: Topic) -> None:
     commit_log_producer = FakeConfluentKafkaProducer()
 
     codec: PassthroughCodec[KafkaPayload] = PassthroughCodec()
-    consumer = KafkaConsumerWithCommitLog(
+    consumer: Consumer[KafkaPayload] = KafkaConsumerWithCommitLog(
         {
             **configuration,
             "auto.offset.reset": "earliest",
